requirements.txt
pillow>=12.0.0
pycryptodome>=3.23.0
pyinstaller>=5.11.2
